requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
